requirements.txt
langchain==0.3.27
python-dotenv==1.1.1
ipykernel==6.30.1
langchain-groq==0.3.7
langchain-google-genai==2.1.10
langchain-community==0.3.29
faiss-cpu==1.12.0
structlog==25.4.0
PyMuPDF==1.26.4
pylint==3.3.8
langchain-core==0.3.75
pytest==8.4.1
streamlit==1.49.1
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
docx2txt==0.9
pypdf==6.0.0
beautifulsoup4==4.13.5
html5lib==1.1
jinja2==3.1.6
lxml==6.0.1
selenium==4.35.0
undetected-chromedriver==3.5.5
langchain_astradb==0.6.1

-e .


